chore(platform): extract shared shell from chat monolith
Create shell.js as the new entry module owning sidebar, theme, status
polling, model switcher bindings, and Escape key handling. chat.js
becomes a chat app module that exports init() and receives shell APIs
via dependency injection. No behavior change — all 58 Playwright E2E
tests pass without modification.

Closes #122
