+++ app/assets/chat.html
@@ -379,6 +379,7 @@
 
   <script src="/assets/vendor/marked.umd.js"></script>
   <script src="/assets/vendor/purify.min.js"></script>
+  <link rel="modulepreload" href="/assets/shell.js">
   <link rel="modulepreload" href="/assets/state.js">
   <link rel="modulepreload" href="/assets/utils.js">
   <link rel="modulepreload" href="/assets/session-manager.js">
@@ -389,6 +390,6 @@
   <link rel="modulepreload" href="/assets/image-handler.js">
   <link rel="modulepreload" href="/assets/settings-ui.js">
   <link rel="modulepreload" href="/assets/chat-engine.js">
-    <script type="module" src="/assets/chat.js"></script>
+    <script type="module" src="/assets/shell.js"></script>
 </body>
 </html>
